:truck: move geometry stuff to Polygon
